Drop obsolete TTF binaries used to store manually generated VTT data
These would need to be regenerated after any changes to the font to be
useful anyway. The data we need to build the fonts is in the TTX files
anyway, and anybody working to re-created it should merge those TTX
dumps into a fresh build of the TTF fonts before proceeding. Keeping
these binaries around doesn't make it easier to contribute to the font,
it only misleads.